refactor: Clear boundaries between CSS and JS
